pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pc15"
version = "15.0.4"
description = "PC15 — whole procedural codec stack (codec, proc, metrics, data, viz, wf)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "PC15 Team" }]
dependencies = [
  'numpy>=1.26,<2.0; python_version < "3.12"',
  'numpy>=2.0; python_version >= "3.12"',
  "Pillow>=10.0",
  "tqdm>=4.66",
  "matplotlib>=3.7",  # for pc15viz; torch intentionally NOT required
]

[tool.setuptools.packages.find]
# Include aggregator src plus each subpackage source dir
where = ["src", "packages/pc15core/src", "packages/pc15proc/src", "packages/pc15codec/src",
         "packages/pc15vq/src", "packages/pc15metrics/src", "packages/pc15data/src",
         "packages/pc15viz/src", "packages/pc15wf/src", "packages/pc15learn/src"]

[tool.setuptools]
include-package-data = true
